+++ src/rich_click/rich_help_rendering.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import inspect
 import re
 from gettext import gettext
-from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, List, Literal, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, List, Literal, Optional, Tuple, Union, overload
 
 import click
 
@@ -32,10 +32,7 @@ from rich_click.rich_parameter import RichParameter
 if TYPE_CHECKING:
     from rich.markdown import Markdown
 
-    from rich_click.rich_help_configuration import (
-        CommandColumnType,
-        OptionColumnType,
-    )
+    from rich_click.rich_help_configuration import CommandColumnType, OptionColumnType, OptionHelpSectionType
     from rich_click.rich_panel import RichCommandPanel, RichOptionPanel
 
 
@@ -128,7 +125,7 @@ def _get_help_text(
     # Get the first paragraph
     first_line = help_text.split("\n\n")[0]
     # Remove single linebreaks
-    if not config.use_markdown and not config.text_markup == "markdown":
+    if not config.text_markup == "markdown":
         if not first_line.startswith("\b"):
             first_line = first_line.replace("\n", " ")
     yield Padding(
@@ -139,7 +136,7 @@ def _get_help_text(
     # Get remaining lines, remove single line breaks and format as dim
     remaining_paragraphs = help_text.split("\n\n")[1:]
 
-    use_markdown = formatter.config.use_markdown or formatter.config.text_markup == "markdown"
+    use_markdown = formatter.config.text_markup == "markdown"
     if formatter.config.text_paragraph_linebreaks is None:
         if use_markdown:
             lb = "\n\n"
@@ -150,12 +147,24 @@ def _get_help_text(
     if len(remaining_paragraphs) > 0:
         if not use_markdown:
             # Remove single linebreaks
-            remaining_paragraphs = [
-                x.replace("\n", " ").strip() if not x.startswith("\b") else "{}\n".format(x.strip("\b\n"))
-                for x in remaining_paragraphs
-            ]
+            _remaining_paragraphs = []
+            for para in remaining_paragraphs:
+                if para.startswith("\b"):
+                    _remaining_paragraphs.append("{}\n".format(para.strip("\b\n")))
+                else:
+                    buf = ""
+                    for p in para.split("\n"):
+                        if any(p.startswith(_) for _ in ["* ", "- ", "    ", "> "]):
+                            if buf:
+                                _remaining_paragraphs.append(buf.strip())
+                                buf = ""
+                            _remaining_paragraphs.append(p.rstrip())
+                        else:
+                            buf += " " + p
+                    if buf:
+                        _remaining_paragraphs.append(buf.strip())
             # Join back together
-            remaining_lines = lb.join(remaining_paragraphs)
+            remaining_lines = lb.join(_remaining_paragraphs)
         else:
             # Join with double linebreaks if markdown
             remaining_lines = lb.join(remaining_paragraphs)
@@ -249,11 +258,66 @@ def _get_parameter_help(
     return formatter.rich_text(help_text, style)
 
 
+@overload
+def _get_parameter_range(
+    param: Union[click.Argument, click.Option, RichParameter],
+    ctx: RichContext,
+    formatter: RichHelpFormatter,
+    mode: Literal["metavar_append"],
+) -> Optional[str]: ...
+
+
+@overload
+def _get_parameter_range(
+    param: Union[click.Argument, click.Option, RichParameter],
+    ctx: RichContext,
+    formatter: RichHelpFormatter,
+    mode: Literal["metavar_column", "help"],
+) -> Optional[Text]: ...
+
+
+def _get_parameter_range(
+    param: Union[click.Argument, click.Option, RichParameter],
+    ctx: RichContext,
+    formatter: RichHelpFormatter,
+    mode: Literal["metavar_append", "metavar_column", "help"],
+) -> Optional[Union[Text, str]]:
+    # Range - from
+    # https://github.com/pallets/click/blob/c63c70dabd3f86ca68678b4f00951f78f52d0270/src/click/core.py#L2698-L2706  # noqa: E501
+    # skip count with default range type
+    if (
+        hasattr(param, "count")
+        and isinstance(param.type, click.types._NumberRangeBase)
+        and not (param.count and param.type.min == 0 and param.type.max is None)
+    ):
+        range_str = param.type._describe_range()
+        if range_str:
+            if mode == "metavar_append":
+                return range_str
+            elif mode == "metavar_column":
+                metavar_str = formatter.config.range_string.format(range_str)
+                return Text.from_markup(metavar_str, style=formatter.config.style_metavar)
+            elif mode == "help":
+                metavar_str = formatter.config.append_range_help_string.format(range_str)
+                return Text.from_markup(
+                    metavar_str,
+                    style=(
+                        formatter.config.style_range_append
+                        if formatter.config.style_range_append is not None
+                        else formatter.config.style_metavar_append
+                    ),
+                )
+            else:
+                raise ValueError("Bad mode selected")
+    return None
+
+
 def _get_parameter_metavar(
     param: Union[click.Argument, click.Option, RichParameter],
     ctx: RichContext,
     formatter: RichHelpFormatter,
     append: bool = True,
+    show_range: bool = False,
 ) -> Optional[Text]:
     metavar_str = param.make_metavar() if CLICK_IS_BEFORE_VERSION_82 else param.make_metavar(ctx)  # type: ignore
     # Do it ourselves if this is a positional argument
@@ -268,6 +332,12 @@ def _get_parameter_metavar(
         metavar_str != "BOOLEAN" and hasattr(param, "is_flag") and not param.is_flag
     ):
         metavar_str = metavar_str.replace("[", "").replace("]", "")
+
+        if show_range:
+            range_txt = _get_parameter_range(param, ctx, formatter, mode="metavar_append")
+            if range_txt:
+                metavar_str += " " + range_txt
+
         return Text.from_markup(
             formatter.config.append_metavars_help_string.format(metavar_str),
             style=formatter.config.style_metavar_append if append else formatter.config.style_metavar,
@@ -280,6 +350,7 @@ def _get_parameter_help_metavar_col(
     param: Union[click.Argument, click.Option, RichParameter],
     ctx: RichContext,
     formatter: RichHelpFormatter,
+    show_range: bool = True,
 ) -> Optional[Text]:
     # Column for a metavar, if we have one
     metavar = Text(style=formatter.config.style_metavar, overflow="fold")
@@ -305,7 +376,7 @@ def _get_parameter_help_metavar_col(
             param.count and param.type.min == 0 and param.type.max is None
         ):
             range_str = param.type._describe_range()
-            if range_str:
+            if show_range and range_str:
                 metavar.append(" " + formatter.config.range_string.format(range_str))
     except AttributeError:
         # click.types._NumberRangeBase is only in Click 8x onwards
@@ -314,7 +385,7 @@ def _get_parameter_help_metavar_col(
     # Highlighter to make [ | ] and <> dim
     class MetavarHighlighter(RegexHighlighter):
         highlights = [
-            r"^(?P<metavar_sep>(\[|<))",
+            r"(^|\s)(?P<metavar_sep>(\[|<))",
             r"(?P<metavar_sep>\|)",
             r"(?P<metavar_sep>(\]|>)$)",
         ]
@@ -441,7 +512,7 @@ def _get_parameter_help_opt(
             return cols[0]
         # Todo - make the text fold at the slashes without adding whitespace.
         #   this is very tricky.
-        t = Text("", overflow="ellipsis").join(cols)
+        t = Text("", overflow="fold").join(cols)
         return t
 
     primary_final = _renderable(primary_cols[:-1])
@@ -552,41 +623,41 @@ def get_help_parameter(
         Columns: A columns element with multiple styled objects (help, default, required)
 
     """
-    config = formatter.config
-    items: List[RenderableType] = []
-
     if TYPE_CHECKING:  # pragma: no cover
         assert isinstance(param.name, str)
 
-    # Get the environment variable first
-    envvar_text = _get_parameter_env_var(param, ctx, formatter)
-    help_text = _get_parameter_help(param, ctx, formatter)
-    deprecated_text = _get_parameter_deprecated(param, ctx, formatter)
-    if formatter.config.append_metavars_help:
-        metavar_text = _get_parameter_metavar(param, ctx, formatter)
-    else:
-        metavar_text = None
-    default_text = _get_parameter_default(param, ctx, formatter)
-    required_text = _get_parameter_required(param, ctx, formatter)
+    section_callbacks: Dict["OptionHelpSectionType", Callable[..., Any]] = {
+        "help": _get_parameter_help,
+        "required": _get_parameter_required,
+        "envvar": _get_parameter_env_var,
+        "default": _get_parameter_default,
+        "metavar": lambda param, ctx, formatter: _get_parameter_metavar(param, ctx, formatter, show_range=True),
+        "metavar_short": lambda param, ctx, formatter: _get_parameter_metavar(param, ctx, formatter, show_range=False),
+        "range": lambda param, ctx, formatter: _get_parameter_range(param, ctx, formatter, mode="help"),
+        "deprecated": _get_parameter_deprecated,
+    }
 
-    if envvar_text is not None and config.option_envvar_first:
-        items.append(envvar_text)
-    if help_text is not None:
-        items.append(help_text)
-    if deprecated_text is not None:
-        items.append(deprecated_text)
-    if metavar_text is not None:
-        items.append(metavar_text)
-    if envvar_text is not None and not config.option_envvar_first:
-        items.append(envvar_text)
-    if default_text is not None:
-        items.append(default_text)
-    if required_text is not None:
-        items.append(required_text)
+    sections: List[Optional[RenderableType]] = []
+    for sec in formatter.config.options_table_help_sections:
+        sections.append(section_callbacks[sec](param, ctx, formatter))
 
     # Use Columns - this allows us to group different renderable types
     # (Text, Markdown) onto a single line.
-    return Columns(items)
+    if formatter.config.text_markup == "markdown" or not all([isinstance(i, Text) or i is None for i in sections]):
+        return Columns([i for i in sections if i])
+    else:
+        # Weird but necessary--
+        # in order to keep things flush, the last column of the table must be of type Columns().
+        # (We are assuming here 'help' is always last, which is not necessarily the case of course.)
+        # In a 2.0 of rich-click we will try to do something nicer than this.
+        # But in 1.x all other solutions would be too breaking.
+        return Columns(
+            [
+                Text(" ", overflow="fold", style=formatter.config.style_option_help).join(
+                    [i for i in sections if i]  # type: ignore[misc]
+                )
+            ]
+        )
 
 
 def get_parameter_rich_table_row(
@@ -662,7 +733,7 @@ def get_parameter_rich_table_row(
 
     _metavar_padded = None
     if any(i in column_types for i in ["opt_all_metavar", "opt_long_metavar"]):
-        _metavar_padded = _get_parameter_metavar(param, ctx, formatter, append=False)
+        _metavar_padded = _get_parameter_metavar(param, ctx, formatter, append=False, show_range=False)
 
     def _opt_all_metavar() -> Optional[RenderableType]:
         if _metavar_padded is None:
@@ -686,6 +757,7 @@ def get_parameter_rich_table_row(
         "opt_secondary": lambda *args, **kwargs: _secondary,
         "opt_all": lambda *args, **kwargs: _all,
         "metavar": _get_parameter_help_metavar_col,
+        "metavar_short": lambda *args, **kwargs: _get_parameter_help_metavar_col(*args, **kwargs, show_range=False),  # type: ignore[misc]
         "opt_all_metavar": lambda *args, **kwargs: _opt_all_metavar(),
         "opt_long_metavar": lambda *args, **kwargs: _opt_long_metavar(),
         "help": lambda *args, **kwargs: (
@@ -809,7 +881,7 @@ def _get_command_help(
 
     paragraphs = inspect.cleandoc(help_text).split("\n\n")
     # Remove single linebreaks
-    if not formatter.config.use_markdown and not paragraphs[0].startswith("\b"):
+    if not formatter.config.text_markup == "markdown" and not paragraphs[0].startswith("\b"):
         paragraphs[0] = paragraphs[0].replace("\n", " ")
     elif paragraphs[0].startswith("\b"):
         paragraphs[0] = paragraphs[0].replace("\b\n", "")