+++ tests/help/test_context_settings.py
@@ -30,8 +30,7 @@ def test_context_settings_help_for_click_8_1_plus(cli_runner: CliRunner, cli: ri
 │ --b        TEXT  This works in all supported click versions. [default: show me always]           │
 │ --c        TEXT  Hide default only in click>=8.1                                                 │
 │ --d        TEXT  Show 'default: (show me in c8+)' in click>=8.0. In click 7, no default is       │
-│                  shown.                                                                          │
-│                  [default: (show me in c8+)]                                                     │
+│                  shown. [default: (show me in c8+)]                                              │
 │ --version        Show the version and exit.                                                      │
 │ --help           Show this message and exit.                                                     │
 ╰──────────────────────────────────────────────────────────────────────────────────────────────────╯
@@ -57,8 +56,7 @@ def test_context_settings_help_for_click_8_0(cli_runner: CliRunner, cli: rich_cl
 │ --b        TEXT  This works in all supported click versions. [default: show me always]           │
 │ --c        TEXT  Hide default only in click>=8.1 [default: show me in old versions]              │
 │ --d        TEXT  Show 'default: (show me in c8+)' in click>=8.0. In click 7, no default is       │
-│                  shown.                                                                          │
-│                  [default: (show me in c8+)]                                                     │
+│                  shown. [default: (show me in c8+)]                                              │
 │ --version        Show the version and exit. [default: False]                                     │
 │ --help           Show this message and exit. [default: False]                                    │
 ╰──────────────────────────────────────────────────────────────────────────────────────────────────╯