+++ CHANGELOG.md
@@ -31,7 +31,9 @@
 - Improved `rich-click` CLI usability: improved patching and improved script detection+execution.
 - `context_settings={"help_to_stderr": True}` support.
 - Added suite of `padding_*` config options to control padding.
-- Added newline control in help text with `text_paragraph_linebreaks` config option.
+  - Better newline handling:
+    - Unordered lists create newlines in help text, e.g. `"\n- foo\n- bar"` or `"\n* foo\n* bar"`
+    - Newline control in help text with `text_paragraph_linebreaks` config option.
 - Added a few `text_*` config options, including notably `text_kwargs` which can control code styles in Markdown.
 
 **Backend (mostly invisible) changes:**